refactor: extract gateway recovery from AgentWorkspace
